+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 New feature(s):
 
 - Add `metadata` command to read and write database metadata.
+- Add `membench` benchmark scores with `scope: RAM` config for the first working set size per operation that exceeds total L3 cache.
 
 ## v0.5.1 (Apr 22, 2026)